DeprecatedGUI: use uintptr_t for SendMessage, drop DWORD typedef (#877)
DeprecatedGUI's SendMessage(int, DWORD, DWORD) round-trips pointers
through DWORD at LVM_GETCURITEM and CBS_ADDSITEM, relying on
sizeof(DWORD) == sizeof(void*). OLX's typedef was `unsigned long`,
which is pointer-sized only on ILP32 and LP64. On LLP64 (Windows
x64) DWORD stays 32 bits while pointers are 64, so the round-trip
silently truncates. A static_assert in main.cpp blocked the
Windows x64 build entirely rather than letting the truncation ship.

DWORD is the wrong type here regardless of platform. By Microsoft's
own definition it is strictly a 32-bit unsigned integer, used for
error codes, sizes, thread IDs, RGBA pixels and flag masks. The
Win32 SendMessage that this code mimics carries pointers through
a different family entirely: WPARAM/LPARAM/LRESULT, which are
UINT_PTR/LONG_PTR — pointer-sized types that Microsoft introduced
specifically for the 32-to-64-bit transition. The very existence
of DWORD_PTR as a separate type from DWORD is implicit confirmation
that DWORD itself is not allowed to grow. The portable C++
equivalent of UINT_PTR is uintptr_t from <cstdint>, which is
defined by the language as "an unsigned integer type capable of
holding a pointer".

In short: this was a wrong-tool-from-the-right-toolbox bug, not a
portability oversight. The original 2002 code targeted only Win32,
where DWORD == WPARAM == LPARAM == void*, and the misuse was
indistinguishable from the correct choice. It only became wrong at
the 64-bit boundary, and on three of four target platforms
`unsigned long` coincidentally tracked `void*`, so the underlying
mistake stayed dormant for 24 years.

Changes:

* Replace DWORD with uintptr_t in CWidget::SendMessage and every
  widget override (CButton, CLabel, CTextbox, CSlider, CScrollbar,
  CCombobox, CListview, CMenu, CImage, CCheckbox, CInputBox,
  CGuiLayout, CGuiSkinnedLayout, CTextButton, CTitleButton, CLine,
  CBox, CBrowser, CProgressbar, CAnimation).
* Update all call sites in src/client/DeprecatedGUI/Menu_*.cpp and
  src/client/CClient_Draw.cpp. (DWORD)0 disambiguation casts become
  (uintptr_t)0.
* Add <cstdint> to include/olx-types.h so uintptr_t is reachable
  wherever DWORD was.
* Remove `typedef unsigned long DWORD;` from olx-types.h.
* Remove the static_assert(sizeof(DWORD) == sizeof(void*)) from
  main.cpp; with uintptr_t the check is vacuous by language
  guarantee.
* Drop now-stale "TODO: 64bit unsafe (pointer cast)" comments at
  CListview.cpp:1724, Menu_Player.cpp:665, CCombobox.cpp:635.

Win32-specific files in src/common/ (Process, SystemFunctions, HTTP,
EventQueue, StackWalker, Debug_DumpCallstack) keep using DWORD via
<windows.h> in actual Win32 API signatures, where it is correct and
unrelated to the misuse this commit fixes.

42 files changed, +192/-194. Verified: Linux x86_64 builds clean.
Unblocks Windows x64.

+++ include/DeprecatedGUI/CCombobox.h
@@ -157,9 +157,9 @@ public:
 	void	Draw(SDL_Surface * bmpDest);
 	void	LoadStyle() {}
 	
-	DWORD SendMessage(int iMsg, DWORD Param1, DWORD Param2);
-	DWORD SendMessage(int iMsg, const std::string& sStr, DWORD Param);
-	DWORD SendMessage(int iMsg, std::string *sStr, DWORD Param);
+	uintptr_t SendMessage(int iMsg, uintptr_t Param1, uintptr_t Param2);
+	uintptr_t SendMessage(int iMsg, const std::string& sStr, uintptr_t Param);
+	uintptr_t SendMessage(int iMsg, std::string *sStr, uintptr_t Param);
 
 	void	setListBackend(const GuiList::Pt& l) { listBackend = l; }
 	/*
