+++ include/DeprecatedGUI/CWidget.h
@@ -155,9 +155,9 @@ public:
 	virtual	void	LoadStyle() = 0;	// Not used anywhere
 	virtual	void	Draw(SDL_Surface * bmpDest) = 0;
 
-	virtual DWORD	SendMessage(int iMsg, DWORD Param1, DWORD Param2) = 0;
-	virtual DWORD	SendMessage(int iMsg, const std::string& sStr, DWORD Param) = 0;
-	virtual DWORD	SendMessage(int iMsg, std::string *sStr, DWORD Param) = 0;
+	virtual uintptr_t	SendMessage(int iMsg, uintptr_t Param1, uintptr_t Param2) = 0;
+	virtual uintptr_t	SendMessage(int iMsg, const std::string& sStr, uintptr_t Param) = 0;
+	virtual uintptr_t	SendMessage(int iMsg, std::string *sStr, uintptr_t Param) = 0;
 	
 	virtual void	ProcessGuiSkinEvent(int iEvent) {};
 };
