+++ src/client/DeprecatedGUI/CListview.cpp
@@ -1687,7 +1687,7 @@ lv_subitem_t *CListview::getCurSubitem(int index)
 
 ///////////////////
 // This widget is send a message
-DWORD CListview::SendMessage(int iMsg, DWORD Param1, DWORD Param2)
+uintptr_t CListview::SendMessage(int iMsg, uintptr_t Param1, uintptr_t Param2)
 {
 	std::string s = "";
 
@@ -1721,7 +1721,7 @@ DWORD CListview::SendMessage(int iMsg, DWORD Param1, DWORD Param2)
         // Return the current item
         case LVM_GETCURITEM:
             if(tSelected)
-                return (DWORD)tSelected; // TODO: 64bit unsafe (pointer cast)
+                return (uintptr_t)tSelected;
             return 0;
 
 		// Set the old-style property
@@ -1742,7 +1742,7 @@ DWORD CListview::SendMessage(int iMsg, DWORD Param1, DWORD Param2)
 	return 0;
 }
 
-DWORD CListview::SendMessage(int iMsg, const std::string& sStr, DWORD Param)
+uintptr_t CListview::SendMessage(int iMsg, const std::string& sStr, uintptr_t Param)
 {
 	switch (iMsg)  {
 
@@ -1764,7 +1764,7 @@ DWORD CListview::SendMessage(int iMsg, const std::string& sStr, DWORD Param)
 }
 
 
-DWORD CListview::SendMessage(int iMsg, std::string *sStr, DWORD Param)
+uintptr_t CListview::SendMessage(int iMsg, std::string *sStr, uintptr_t Param)
 {
 	switch (iMsg)  {
