+++ src/client/DeprecatedGUI/CTextbox.cpp
@@ -705,13 +705,13 @@ void CTextbox::setText(const std::string& buf) {
 
 ///////////////////
 // This widget is send a message
-DWORD CTextbox::SendMessage(int iMsg, DWORD Param1, DWORD Param2) {
+uintptr_t CTextbox::SendMessage(int iMsg, uintptr_t Param1, uintptr_t Param2) {
 
 	switch(iMsg) {
 
 		// Get the text length
 		case TXM_GETTEXTLENGTH:
-			return (DWORD)Utf8StringSize(sText);
+			return (uintptr_t)Utf8StringSize(sText);
 			break;
 
 		// Set some flags
@@ -732,7 +732,7 @@ DWORD CTextbox::SendMessage(int iMsg, DWORD Param1, DWORD Param2) {
 	return 0;
 }
 
-DWORD CTextbox::SendMessage(int iMsg, const std::string& sStr, DWORD Param)
+uintptr_t CTextbox::SendMessage(int iMsg, const std::string& sStr, uintptr_t Param)
 {
 	switch (iMsg)  {
 
@@ -746,7 +746,7 @@ DWORD CTextbox::SendMessage(int iMsg, const std::string& sStr, DWORD Param)
 }
 
 
-DWORD CTextbox::SendMessage(int iMsg, std::string *sStr, DWORD Param)
+uintptr_t CTextbox::SendMessage(int iMsg, std::string *sStr, uintptr_t Param)
 {
 	switch (iMsg)  {
